Create Control flow: Else statement
Calvin Coolidge’s Cool College has another request for you. They want you to add an additional check to the graduation_reqs function. If a student is failing to meet both graduation requirements, they want the function to return:
"You do not meet the GPA or the credit requirement for graduation."
Use an else statement to add this to your function.