Create Challenges
a:greater number:
The function should return True if num is greater than or equal to lower and less than or equal to upper. Otherwise, return False.
b:identical names:
:If our names are identical, return True. Otherwise, return False.
c:Always False:
Using an if statement, your variable num, and the operators >, and <, make it so your function will return False no matter what number is stored in num.
An if statement that is always false is called a contradiction. You will rarely want to do this while programming, but it is important to realize it is possible to do this.
d:Movie rating:
If rating is less than or equal to 5, return "Avoid at all costs!". If rating is between 5 and 9, return "This one was fun.". If rating is 9 or above, return "Outstanding!"
e:Max Number:
The function should return the largest of these three numbers. If any of two numbers tie as the largest, you should return "It's a tie!".